Feat: 테스트코드 폴더 생성, supertest를 위한 서버 listening 분리
