+++ package.json
@@ -25,7 +25,7 @@
   "version": "1.0.0",
   "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -39,6 +39,8 @@
   },
   "homepage": "https://github.com/Hanghae-Realwork/back_end#readme",
   "devDependencies": {
-    "sequelize-cli": "^6.4.1"
+    "jest": "^28.1.3",
+    "sequelize-cli": "^6.4.1",
+    "supertest": "^6.2.4"
   }
 }
